Fixed remaining warnings primarily introduced by C++ language upgrade.
Some minor ones still remain (will fix at a later point), but for now the main source of compiler errors relates to missing third-party libraries.

+++ Code/Commando/WOLGameInfo.cpp
@@ -314,7 +314,7 @@ void WOLGameInfo::ImportFromChannel(const RefPtr<ChannelData>& channel)
 		mIsDataValid = false;
 		}
 
-	if (version != cNetwork::Get_Exe_Key()) {
+	if (version != (unsigned long)cNetwork::Get_Exe_Key()) {
 		mIsDataValid = false;
 		return;
 	}
